Moved UpdateRelatedSeries from controller to SeriesService.cs
