+++ API/Services/SeriesService.cs
@@ -5,12 +5,14 @@ using System.Linq;
 using System.Threading.Tasks;
 using API.Comparators;
 using API.Data;
+using API.Data.Repositories;
 using API.DTOs;
 using API.DTOs.CollectionTags;
 using API.DTOs.Metadata;
 using API.DTOs.SeriesDetail;
 using API.Entities;
 using API.Entities.Enums;
+using API.Entities.Metadata;
 using API.Helpers;
 using API.SignalR;
 using Microsoft.AspNetCore.Mvc;
@@ -26,6 +28,7 @@ public interface ISeriesService
     Task<bool> UpdateRating(AppUser user, UpdateSeriesRatingDto updateSeriesRatingDto);
     Task<bool> DeleteMultipleSeries(IList<int> seriesIds);
 
+    Task<bool> UpdateRelatedSeries(UpdateRelatedSeriesDto dto);
 }
 
 public class SeriesService : ISeriesService
@@ -608,5 +611,63 @@ public class SeriesService : ISeriesService
             LibraryType.Book => "Book",
             _ => "Chapter"
         };
+    }
+
+    public async Task<bool> UpdateRelatedSeries(UpdateRelatedSeriesDto dto)
+    {
+        var series = await _unitOfWork.SeriesRepository.GetSeriesByIdAsync(dto.SeriesId, SeriesIncludes.Related);
+
+        UpdateRelationForKind(dto.Adaptations, series.Relations.Where(r => r.RelationKind == RelationKind.Adaptation).ToList(), series, RelationKind.Adaptation);
+        UpdateRelationForKind(dto.Characters, series.Relations.Where(r => r.RelationKind == RelationKind.Character).ToList(), series, RelationKind.Character);
+        UpdateRelationForKind(dto.Contains, series.Relations.Where(r => r.RelationKind == RelationKind.Contains).ToList(), series, RelationKind.Contains);
+        UpdateRelationForKind(dto.Others, series.Relations.Where(r => r.RelationKind == RelationKind.Other).ToList(), series, RelationKind.Other);
+        UpdateRelationForKind(dto.SideStories, series.Relations.Where(r => r.RelationKind == RelationKind.SideStory).ToList(), series, RelationKind.SideStory);
+        UpdateRelationForKind(dto.SpinOffs, series.Relations.Where(r => r.RelationKind == RelationKind.SpinOff).ToList(), series, RelationKind.SpinOff);
+        UpdateRelationForKind(dto.AlternativeSettings, series.Relations.Where(r => r.RelationKind == RelationKind.AlternativeSetting).ToList(), series, RelationKind.AlternativeSetting);
+        UpdateRelationForKind(dto.AlternativeVersions, series.Relations.Where(r => r.RelationKind == RelationKind.AlternativeVersion).ToList(), series, RelationKind.AlternativeVersion);
+        UpdateRelationForKind(dto.Doujinshis, series.Relations.Where(r => r.RelationKind == RelationKind.Doujinshi).ToList(), series, RelationKind.Doujinshi);
+        UpdateRelationForKind(dto.Prequels, series.Relations.Where(r => r.RelationKind == RelationKind.Prequel).ToList(), series, RelationKind.Prequel);
+        UpdateRelationForKind(dto.Sequels, series.Relations.Where(r => r.RelationKind == RelationKind.Sequel).ToList(), series, RelationKind.Sequel);
+
+        if (!_unitOfWork.HasChanges()) return true;
+        if (await _unitOfWork.CommitAsync()) return true;
+
+
+        return false;
+    }
+
+
+    /// <summary>
+    ///
+    /// </summary>
+    /// <param name="dtoTargetSeriesIds"></param>
+    /// <param name="adaptations"></param>
+    /// <param name="series"></param>
+    /// <param name="kind"></param>
+    private void UpdateRelationForKind(ICollection<int> dtoTargetSeriesIds, IEnumerable<SeriesRelation> adaptations, Series series, RelationKind kind)
+    {
+        foreach (var adaptation in adaptations.Where(adaptation => !dtoTargetSeriesIds.Contains(adaptation.TargetSeriesId)))
+        {
+            // If the seriesId isn't in dto, it means we've removed or reclassified
+            series.Relations.Remove(adaptation);
+        }
+
+        // At this point, we only have things to add
+        foreach (var targetSeriesId in dtoTargetSeriesIds)
+        {
+            // This ensures we don't allow any duplicates to be added
+            if (series.Relations.SingleOrDefault(r =>
+                    r.RelationKind == kind && r.TargetSeriesId == targetSeriesId) !=
+                null) continue;
+
+            series.Relations.Add(new SeriesRelation()
+            {
+                Series = series,
+                SeriesId = series.Id,
+                TargetSeriesId = targetSeriesId,
+                RelationKind = kind
+            });
+            _unitOfWork.SeriesRepository.Update(series);
+        }
     }
 }
